Move core to private dir
