+++ include/aws/crt/mqtt/private/MqttConnectionCore.h
